requirements.txt
streamlit
numpy
pandas
scikit-learn
streamlit-drawable-canvas
joblib
cv2-python